multi_agents/requirements.txt
langgraph
gpt_researcher
langgraph-cli
python-dotenv
weasyprint
json5
loguru
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0